TechDraw: changed CosmeticVertex point picker dashed circle to filled dot. (#20095)
* techdraw: changed the cosmetic vertex shape

* add back the call to prepareGeometryChange

* add getVertexSize and getLineWidth methods to QGIViewPart

* use doubles for sizes

* update setRadius in QGMarker to use double instead of float

+++ src/Mod/TechDraw/Gui/QGIVertex.cpp
@@ -49,7 +49,7 @@ QGIVertex::QGIVertex(int index) :
     setRadius(m_radius);
 }
 
-void QGIVertex::setRadius(float r)
+void QGIVertex::setRadius(double r)
 {
     m_radius = r;
     QPainterPath p;
